pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "setlist-sync"
version = "0.1.0"
description = "Match streaming playlists to your DJ library"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "rapidfuzz>=3.0",
    "spotifyscraper>=2.0",
    "python-dotenv>=1.0",
    "mutagen>=1.47",
    "pyrekordbox>=0.4",
]

[project.scripts]
setlist-sync = "setlist_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
